+++ Sources/NetworkUtilities/Params.swift
@@ -1,6 +1,6 @@
 //
 //  DateFormatter.swift
-//  EVPulse
+//  
 //
 //  Created by Nikolas Konstantakopoulos on 14/2/22.
 //